feat: 상품 추가 화면에서 Cancel 버튼과 Done 버튼을 눌렀을 때의 기능 구현

+++ OpenMarket/OpenMarket/Model/RegistProduct/ProductParams.swift
@@ -10,9 +10,9 @@ import Foundation
 struct ProductParams: Codable {
     let name: String
     let descriptions: String
-    let price: Int
+    let price: Double
     let currency: Currency
-    let discountedPrice: Int
+    let discountedPrice: Double
     let stock: Int
     let secret: String
